requirements.txt
Flask
gunicorn
google-cloud-datastore
openai==0.28.0
pymupdf
python-docx
flask-cors


